+++ src/main/java/enjoytrip/member/service/MemberService.java
@@ -1,7 +1,11 @@
 package enjoytrip.member.service;
 
 import enjoytrip.member.domain.Member;
-import enjoytrip.member.dto.*;
+import enjoytrip.member.dto.request.MemberSaveRequest;
+import enjoytrip.member.dto.request.MemberUpdateRequest;
+import enjoytrip.member.dto.response.MemberFindResponse;
+import enjoytrip.member.dto.response.MemberSaveResponse;
+import enjoytrip.member.dto.response.MemberUpdateResponse;
 import enjoytrip.member.exception.MemberNotFoundException;
 import enjoytrip.member.repository.MemberRepository;
 import lombok.AllArgsConstructor;
